# src/engine/server/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  util.Config:
    properties:
      appPlugin:
        type: string
      appPluginParameters:
        type: object
      appQuiesceCmd:
        type: string
      appUnquiesceCmd:
        type: string
      archiveCreateCmd:
        type: string
      archiveDeleteCmd:
        type: string
      archivePlugin:
        type: string
      autoDiscovery:
        type: boolean
      backupCreateCmd:
        type: string
      backupDeleteCmd:
        type: string
      backupPolicy:
        type: string
      backupRetention:
        type: integer
      backupRetentions:
        items:
          type: BackupRetention
        type: array
      configName:
        type: string
      jobRetention:
        type: integer
      postAppQuiesceCmd:
        type: string
      postAppRestoreCmd:
        type: string
      postAppUnquiesceCmd:
        type: string
      preAppQuiesceCmd:
        type: string
      preAppRestoreCmd:
        type: string
      preAppUnquiesceCmd:
        type: string
      profileName:
        type: string
      restoreCmd:
        type: string
      selectedWorkflowId:
        type: integer
      sendTrapErrorCmd:
        type: string
      sendTrapSuccessCmd:
        type: string
      storagePlugin:
        type: string
      storagePluginParameters:
        type: object
      workflowId:
        type: string
    type: object
...
  util.Jobs:
    properties:
      backup:
        items:
          type: Job
        type: array
      result:
        $ref: '#/definitions/util.Result'
        type: object
    type: object
...
  util.Result:
    properties:
      code:
        type: integer
      data:
        items:
          type: string
        type: array
      messages:
        items:
          type: Message
        type: array
    type: object
...
host: '{{.Host}}'
info:
  contact:
    email: keith.tenzer@gmail.com
    name: Keith Tenzer
    url: http://www.keithtenzer.com
  description: |-
    APIs for managing Fossul workflows, jobs, profile, and configs
    JSON API definition can be retrieved at <a href="/api/v1/swagger/doc.json">/api/v1/swagger/doc.json</a>
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fossul Framework Server API
  version: "1.0"
paths:
...
  /getJobs/{profileName}/{configName}:
    get:
      consumes:
      - application/json
      description: Get jobs (workflows) that have executed for a profile/config
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Jobs'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
  /sendTrapErrorCmd:
    post:
      consumes:
      - application/json
      description: Execute command after failed workflow execution
      parameters:
      - description: config struct
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/util.Config'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /sendTrapSuccessCmd:
    post:
      consumes:
      - application/json
      description: Execute command after successfull workflow execution
      parameters:
      - description: config struct
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/util.Config'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
  /status:
    get:
      consumes:
      - application/json
      description: Status and version information for the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
swagger: "2.0"